requirements.txt
discord.py>=2.3.2
aiohttp>=3.8.0
requests==2.31.0